+++ lib/osf-components/addon/components/add-your-research/component.ts
@@ -1,11 +1,12 @@
-import { tagName } from '@ember-decorators/component';
 import Component from '@ember/component';
 import { layout } from 'ember-osf-web/decorators/component';
+
+import defaultTo from 'ember-osf-web/utils/default-to';
 
 import styles from './styles';
 import template from './template';
 
-@tagName('')
 @layout(template, styles)
 export default class AddYourResearch extends Component {
+    version: string = defaultTo(this.version, 'versionA');
 }