+++ lib/osf-components/addon/components/discover-research/component.ts
@@ -2,12 +2,14 @@ import { action } from '@ember-decorators/object';
 import Component from '@ember/component';
 
 import { layout, requiredAction } from 'ember-osf-web/decorators/component';
+import defaultTo from 'ember-osf-web/utils/default-to';
 import styles from './styles';
 import template from './template';
 
 @layout(template, styles)
 export default class DiscoverResearch extends Component {
     value: string = '';
+    version: string = defaultTo(this.version, 'versionA');
     @requiredAction onSearch!: (query: string) => void;
 
     @action